requirements.txt
asgiref==3.8.1
certifi==2025.11.12
charset-normalizer==3.4.4
dj-database-url==2.3.0
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
Faker==33.3.0
GDAL==3.6.2
gunicorn==23.0.0
idna==3.11
packaging==24.2
pillow==11.1.0
pip==24.3.1
psycopg==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.6.2
whitenoise==6.8.2
